new network directory
