+++ include/mdns_client.h
@@ -1,0 +1,6 @@
+#include <Arduino.h>
+
+// mdns.h ではないのは、 dunk mdnsのヘッダファイルが mdns.h なため
+
+void mdns_setup();
+void mdns_loop();
