+++ include/wifi.h
@@ -3,5 +3,3 @@ void make_sure_wifi_connected();
 void start_wifi_access_point();
 
 String get_wifi_ip_addr();
-
-bool start_mdns(String name);
