+++ platformio.ini
@@ -8,46 +8,6 @@
 ; Please visit documentation for the other options and examples
 ; https://docs.platformio.org/page/projectconf.html
 
-; [env:esp32dev]
-; platform = espressif32
-; board = esp32dev
-; framework = arduino
-; board_build.mcu = esp32
-; monitor_speed = 115200
-; board_build.f_cpu = 80000000L
-; lib_deps = 
-; 	joba-1/Joba_Tsl2561@^2.0.10
-; 	squix78/ESP8266_SSD1306@^4.1.0
-; 	256dpi/MQTT@^2.4.7
-; 	wifwaf/MH-Z19@^1.5.2
-; 	sodaqmoja/Sodaq_LPS22HB@^1.0.0
-; 	ottowinter/ESPAsyncWebServer-esphome@^1.2.7
-; 	lorol/LittleFS_esp32@^1.0.0
-; 	seeed-studio/Grove - Barometer Sensor BME280@^1.0.2
-; 	bodmer/TFT_eSPI@^2.3.41
-; build_flags = 
-; 	-Os
-; 	-DCORE_DEBUG_LEVEL=2
-; 	-DUSER_SETUP_LOADED=1
-; 	-DST7789_DRIVER=1
-; 	-DTFT_WIDTH=135
-; 	-DTFT_HEIGHT=240
-; 	-DTFT_MOSI=19
-; 	-DTFT_SCLK=18
-; 	-DTFT_CS=5
-; 	-DTFT_DC=16
-; 	-DTFT_RST=23
-; 	-DTFT_BL=4
-; 	-DSPI_FREQUENCY=40000000
-; 	-DLOAD_GLCD=1
-; 	-DCGRAM_OFFSET=1
-; 	-DLOAD_FONT2=1
-; 	-DLOAD_FONT4=1
-; 	-DLOAD_FONT6=1
-; 	-DLOAD_FONT7=1
-; 	-DLOAD_GFXFF=1
-; 	-DSMOOTH_FONT=1
-
 [env:esp12e]
 platform = espressif8266
 board = esp12e
@@ -61,5 +21,6 @@ lib_deps =
 	sodaqmoja/Sodaq_LPS22HB@^1.0.0
 	ottowinter/ESPAsyncWebServer-esphome@^1.2.7
 	seeed-studio/Grove - Barometer Sensor BME280@^1.0.2
+	mrhornsby/MDNS@^2.0.0
 build_flags = 
 	-DCORE_DEBUG_LEVEL=2
