+++ src/network/wifi.cpp
@@ -7,13 +7,6 @@
 #include <ESP8266WiFi.h>
 #endif
 
-// mDNS
-#ifdef ESP32
-#include <ESPmDNS.h>
-#elif defined(ESP8266)
-#include <ESP8266mDNS.h>
-#endif
-
 #include "log.h"
 #include "global.h"
 
@@ -124,14 +117,3 @@ String get_wifi_ip_addr() {
 }
 
 
-bool start_mdns(String name) {
-  char n[name.length() + 1];
-  name.toCharArray(n, sizeof n);
-
-  if (!MDNS.begin(n)) {
-    mdnslog("Error setting up MDNS responder!");
-    return false;
-  }
-  mdnslog("mDNS responder started");
-  return true;
-}
